+++ version/changelog.md
@@ -8,6 +8,7 @@
 - ENTER executes an executable
 - display own icons for files depending on extension
 - typing text will execute commands in cmdline
+- delete key should also delete file like F8
 
 ### v1.1.0 (26.01.2026)